### 2018-08-05 #874 @whedon
[PRE REVIEW]: nestcheck: error analysis, diagnostic tests and plots for nested sampling calculations
**Submitting author:** @ejhigson (<a href="http://orcid.org/0000-0001-8383-4614">Edward Higson</a>)
**Repository:** <a href="https://github.com/ejhigson/nestcheck" target ="_blank">https://github.com/ejhigson/nestcheck</a>
**Version:** v0.1.5
