### 2018-08-05 3ed06d5
removing typo in paper reference key
