+++ docs/paper/paper.md
@@ -26,7 +26,7 @@ Due to the distinctive manner in which the nested sampling algorithm explores th
 As a result, posterior inferences and estimates of their associated uncertainties require methods specific to nested sampling.
 
 ``nestcheck`` is a Python package for analysing samples produced by nested sampling, and estimating uncertainty on posterior inferences.
-Most importantly, ``nestcheck`` contains fast and well-tested implementations of the error analysis methods introduced in [@Higson2017a] and the diagnostic tests and plots described in [@Higson:2018a].
+Most importantly, ``nestcheck`` contains fast and well-tested implementations of the error analysis methods introduced in [@Higson2017a] and the diagnostic tests and plots described in [@Higson2018a].
 The code has been profiled for computational efficiency and uses fast ``numpy`` functions and parallelisation with ``concurrent.futures``.
 The diagnostic plots make use of the ``matplotlib`` [@Hunter2007] and ``fgivenx`` [@zenodofgivenx] packages.
 
@@ -34,7 +34,7 @@ The diagnostic plots make use of the ``matplotlib`` [@Hunter2007] and ``fgivenx`
 ``nestcheck`` is also compatible with samples produced by the dynamic nested sampling algorithm [@Higson2017b], and its functions for storing and manipulating nested sampling output are used by the ``dyPolyChord`` [@zenododypolychord] and ``perfectns`` [@zenodoperfectns] dynamic nested sampling packages.
 
 ``nestcheck`` is designed to allow nested sampling software users to quickly calculate results and uncertainty estimates, as well as to apply diagnostics for checking their software has explored the posterior correctly.
-It was used for the diagnostic tests and plots in [@Higson:2018a], and for error analysis in [@Higson2017b] and [@Higson2018b].
+It was used for the diagnostic tests and plots in [@Higson2018a], and for error analysis in [@Higson2017b] and [@Higson2018b].
 An earlier version of the code was used in the analysis of black hole mergers in [@Chua2018].
 The source code for ``nestcheck`` has been archived to Zenodo [@zenodonestcheck].
 
