### 2018-08-05 2a0f974
minor edits to text and bibtex
