### 2018-08-11 47f6554
moving efficiency gain calculation to helper function
