### 2018-08-15 e96dc9b
improving error handling and docs
