### 2018-08-16 271663b
changing calls to fgivenx for compatibility with its new version, and updating versions and dependencies
