### 2018-08-16 5c149a0
updating error message for old PolyChord bug
