### 2018-08-16 f4befc4
changing PolyChord version requirement to avoid bug in v1.13
