### 2018-08-16 #874 @ejhigson
Perhaps @mattpitkin or @ziotom78 would be suitable reviewers?
