### 2018-08-19 #874 @nespinoza
Hi @ejhigson; would love to but this week I'm in a conference, and next on vacation. I'll be available from September 3rd and on --- if this works for you, I could make my report for that week.

Best,

Néstor
