### 2018-08-22 0d706c3
added thread decomposition dealing with points being born and dying on same contour. Gives valid decomposition even with logl=const and many duplicates
