### 2018-08-27 6225268
minor tidying and improving error message
