### 2018-08-27 980d46e
specifically handle ValueError in reading PolyChord .stats nlike with rest of file working
