### 2018-08-27 6c3a979
minor changes to docstring
