### 2018-08-27 #874 @ejhigson
Hi @arfon - maybe mattpitkin and one of ziotom78 and nespinoza would be a good pair of reviewers? I'm happy to wait until the latter gets back from vacation. Thanks very much!
