### 2018-08-27 23036cd
edited error raising to ensure compatibility with python2
