### 2018-08-28 #874 @arfon
@ziotom78, @mattpitkin - thanks again for agreeing to review this submission. See you over in #916 for the actual review.
