### 2018-09-02 1b65a46
minor doc update
