### 2018-09-04 1e493f1
adding random assignment of any points left without a thread label to different threads
