### 2018-09-04 78cf10f
adding travis email notifications
