+++ .travis.yml
@@ -16,3 +16,6 @@ script:
  - nosetests --with-coverage --cover-erase --cover-package=nestcheck
 after_success:
  - coveralls
+notifications:
+  email:
+    - e.higson@mrao.cam.ac.uk
