### 2018-09-07 ae673e4
minor update to error message
