### 2018-09-10 #916 @mattpitkin
@ejhigson Sorry for the delay in starting the review, but I'm getting started now.

First, some easy fixes. In the references could you:

* add the DOI for the Matplotlib paper https://doi.org/10.1109/MCSE.2007.55
* add the arXiv URL for the "Importance Nested Sampling..." paper https://arxiv.org/abs/1306.2144
