### 2018-09-10 #916 @mattpitkin
Just a small request and query for the "Quickstart demo".

In the "**Calculating errors due to implementation-specific effects**" [section](https://nestcheck.readthedocs.io/en/latest/demos/quickstart_demo.html#Calculating-errors-due-to-implementation-specific-effects) could you state explicitly that this requires you to have multiple run (I know this is kind of covered in the next section, but it would be useful to have it).

In the "**Tests for implementation specific effects using only 2 nested sampling runs**" [section](https://nestcheck.readthedocs.io/en/latest/demos/quickstart_demo.html#Tests-for-implementation-specific-effects-using-only-2-nested-sampling-runs) should the line:

> As expected, the p-values are not particularly low nor are the KS distances particularly high - indicating there are significant implementation-specific effects for the simple 2-dimensional Gaussian

actually say

> As expected, the p-values are not particularly low nor are the KS distances particularly high - indicating there are **no** significant implementation-specific effects for the simple 2-dimensional Gaussian
