### 2018-09-10 #916 @mattpitkin
Would you be able to add some more information to the documentation about how to go about allowing samples from different nested sampling codes to be read in to `nestcheck`? I.e., what should a code output to allow you to use it with `nestcheck` through a custom `process_XXXX_runs` function? I see from the [`process_dynesty_runs`](https://github.com/ejhigson/nestcheck/blob/master/nestcheck/data_processing.py#L254) source that it seems to need values of the log likelihood, the actual sample values, the number of live points, and the sample ID - is the sample ID the order with which that sample was used for the nested sample integral (i.e., when that point "died")? What is a "thread" in this context?
