### 2018-09-10 #916 @ziotom78
@arfon, I have downloaded the two papers cited by the package's authors and have discovered that they used it for work related to the Planck project, to whom I have belonged for a long time. I do not think there is a clear conflict of interest, as I have never interacted with any of the authors about the topics covered here. However, the JOSS Code of Conduct states clearly that I need to report this.
