### 2018-09-11 68c9a1e
updating paper references
