+++ docs/paper/paper.bib
@@ -17,7 +17,8 @@ number = {3},
 pages = {90--95},
 title = {{Matplotlib: A 2D graphics environment}},
 volume = {9},
-year = {2007}
+year = {2007},
+doi = {https://doi.org/10.1109/MCSE.2007.55}
 }
 @article{Feroz2013,
 archivePrefix = {arXiv},
@@ -27,6 +28,7 @@ eprint = {1306.2144},
 journal = {arXiv preprint arXiv:1306.2144},
 keywords = {bayesian methods,data analysis,model selection,monte carlo methods},
 title = {{Importance Nested Sampling and the MultiNest Algorithm}},
+url = {https://arxiv.org/abs/1306.2144},
 year = {2013}
 }
 @article{Feroz2009,
@@ -107,22 +109,28 @@ url = {http://dx.doi.org/10.1093/mnras/sty1079},
 volume = {478},
 year = {2018}
 }
-@misc{zenodofgivenx,
+@article{Handley2018fgivenx,
 author = {Handley, Will},
-doi = {10.5281/zenodo.1310715},
-title = {fgivenx},
+doi = {10.21105/joss.00849},
+issn = {2475-9066},
+journal = {Journal of Open Source Software},
+number = {28},
+pages = {849},
+title = {{fgivenx: A Python package for functional posterior plotting}},
+url = {http://joss.theoj.org/papers/10.21105/joss.00849},
+volume = {3},
 year = {2018}
 }
-@article{zenododypolychord,
+@article{Higson2018dypolychord,
 author = {Higson, Edward},
 doi = {10.5281/zenodo.1328175},
 title = {{dyPolyChord: dynamic nested sampling with PolyChord}},
 year = {2018}
 }
-@article{zenodoperfectns,
+@article{Higson2018perfectns,
 author = {Higson, Edward},
 doi = {10.5281/zenodo.1327591},
-title = {perfectns: dynamic and standard nested sampling for spherically symmetric likelihoods and priors},
+title = {perfectns: perfect dynamic and standard nested sampling for spherically symmetric likelihoods and priors},
 year = {2018}
 }
 @misc{zenodonestcheck,
