+++ docs/paper/paper.md
@@ -28,10 +28,10 @@ As a result, posterior inferences and estimates of their associated uncertaintie
 ``nestcheck`` is a Python package for analysing samples produced by nested sampling, and estimating uncertainty on posterior inferences.
 Most importantly, ``nestcheck`` contains fast and well-tested implementations of the error analysis methods introduced in [@Higson2017a] and the diagnostic tests and plots described in [@Higson2018a].
 The code has been profiled for computational efficiency and uses fast ``numpy`` functions and parallelisation with ``concurrent.futures``.
-The diagnostic plots make use of the ``matplotlib`` [@Hunter2007] and ``fgivenx`` [@zenodofgivenx] packages.
+The diagnostic plots make use of the ``matplotlib`` [@Hunter2007] and ``fgivenx`` [@Handley2018fgivenx] packages.
 
 ``nestcheck`` can analyse samples from the popular ``MultiNest`` [@Feroz2008; @Feroz2009; @Feroz2013] and ``PolyChord`` [@Handley2015a; @Handley2015b] packages, and functions for loading samples from other software packages with different formats can easily be added.
-``nestcheck`` is also compatible with samples produced by the dynamic nested sampling algorithm [@Higson2017b], and its functions for storing and manipulating nested sampling output are used by the ``dyPolyChord`` [@zenododypolychord] and ``perfectns`` [@zenodoperfectns] dynamic nested sampling packages.
+``nestcheck`` is also compatible with samples produced by the dynamic nested sampling algorithm [@Higson2017b], and its functions for storing and manipulating nested sampling output are used by the ``dyPolyChord`` [@Higson2018dypolychord] and ``perfectns`` [@Higson2018perfectns] dynamic nested sampling packages.
 
 ``nestcheck`` is designed to allow nested sampling software users to quickly calculate results and uncertainty estimates, as well as to apply diagnostics for checking their software has explored the posterior correctly.
 It was used for the diagnostic tests and plots in [@Higson2018a], and for error analysis in [@Higson2017b] and [@Higson2018b].
