### 2018-09-11 76d76bd
minor doc updates
