### 2018-09-11 7b9b8e7
expanding function documentation of estimaters.py
