+++ nestcheck/estimators.py
@@ -1,14 +1,17 @@
 #!/usr/bin/env python
 """
 Functions for estimating quantities from nested sampling runs.
-Each estimator function should have arguments
+Each estimator function should have arguments:
 
 .. code-block:: python
 
     def estimator_func(self, ns_run, logw=None, simulate=False):
         ...
 
-The `logw` argument allows the log weights for the points in the run to be
+Any additional arguments required for the function should be keyword
+arguments.
+
+The ``logw`` argument allows the log weights for the points in the run to be
 provided - this is useful if many estimators are being calculated from
 the same run as it allows `logw` to only be calculated once. If it is not
 specified, `logw` is calculated from the run when required.
@@ -81,9 +84,21 @@ def get_latex_name(func_in, **kwargs):
 # ----------
 
 def count_samples(ns_run, **kwargs):
-    """Number of samples in run. Unlike most estimators this does not require
-    log weights, but for convenience will not throw an error if they are
-    specified."""
+    """Number of samples in run.
+
+    Unlike most estimators this does not require log weights, but for
+    convenience will not throw an error if they are specified.
+
+    Parameters
+    ----------
+    ns_run: dict
+        Nested sampling run dict (see the ``data_processing`` module
+        docstring for more details).
+
+    Returns
+    -------
+    int
+    """
     kwargs.pop('logw', None)
     kwargs.pop('simulate', None)
     if kwargs:
@@ -92,14 +107,46 @@ def count_samples(ns_run, **kwargs):
 
 
 def logz(ns_run, logw=None, simulate=False):
-    """Natural log of Bayesian evidence."""
+    """Natural log of Bayesian evidence :math:`\log \mathcal{Z}`.
+
+    Parameters
+    ----------
+    ns_run: dict
+        Nested sampling run dict (see the data_processing module
+        docstring for more details).
+    logw: None or 1d numpy array, optional
+        Log weights of samples.
+    simulate: bool, optional
+        Passed to ns_run_utils.get_logw if logw needs to be
+        calculated.
+
+    Returns
+    -------
+    float
+    """
     if logw is None:
         logw = nestcheck.ns_run_utils.get_logw(ns_run, simulate=simulate)
     return scipy.special.logsumexp(logw)
 
 
 def evidence(ns_run, logw=None, simulate=False):
-    """Bayesian evidence."""
+    """Bayesian evidence :math:`\log \mathcal{Z}`.
+
+    Parameters
+    ----------
+    ns_run: dict
+        Nested sampling run dict (see the data_processing module
+        docstring for more details).
+    logw: None or 1d numpy array, optional
+        Log weights of samples.
+    simulate: bool, optional
+        Passed to ns_run_utils.get_logw if logw needs to be
+        calculated.
+
+    Returns
+    -------
+    float
+    """
     if logw is None:
         logw = nestcheck.ns_run_utils.get_logw(ns_run, simulate=simulate)
     return np.exp(scipy.special.logsumexp(logw))
@@ -107,12 +154,30 @@ def evidence(ns_run, logw=None, simulate=False):
 
 def param_mean(ns_run, logw=None, simulate=False, param_ind=0,
                handle_indexerror=False):
-    """
-    Mean of a single parameter (single component of theta).
+    """Mean of a single parameter (single component of theta).
 
-    handle_indexerror means the function returns nan rather than raising an
-    IndexError if param_ind >= ndim. This is useful when applying the same
-    list of estimators to data sets of different dimensions.
+    Parameters
+    ----------
+    ns_run: dict
+        Nested sampling run dict (see the data_processing module
+        docstring for more details).
+    logw: None or 1d numpy array, optional
+        Log weights of samples.
+    simulate: bool, optional
[…]
