### 2018-09-11 4be90cc
moving latex names function to bottom of file
