### 2018-09-11 6f4cc28
minor tidying of docs and messages
