### 2018-09-11 c698120
improving error_analysis documentation
