+++ nestcheck/error_analysis.py
@@ -89,7 +89,7 @@ def run_std_bootstrap(ns_run, estimator_list, **kwargs):
 
     For more details about bootstrap resampling for estimating sampling
     errors see 'Sampling errors in nested sampling parameter estimation'
-    (Higson et al. 2017).
+    (Higson et al. 2018).
 
     Parameters
     ----------
@@ -123,7 +123,7 @@ def run_bootstrap_values(ns_run, estimator_list, **kwargs):
 
     For more details about bootstrap resampling for estimating sampling
     errors see 'Sampling errors in nested sampling parameter estimation'
-    (Higson et al. 2017).
+    (Higson et al. 2018).
 
     Parameters
     ----------
@@ -187,7 +187,7 @@ def run_ci_bootstrap(ns_run, estimator_list, **kwargs):
 
     For more details about bootstrap resampling for estimating sampling
     errors see 'Sampling errors in nested sampling parameter estimation'
-    (Higson et al. 2017).
+    (Higson et al. 2018).
 
     Parameters
     ----------
@@ -234,7 +234,7 @@ def run_std_simulate(ns_run, estimator_list, n_simulate=None):
 
     For more details about the simulated weights method for estimating sampling
     errors see 'Sampling errors in nested sampling parameter estimation'
-    (Higson et al. 2017).
+    (Higson et al. 2018).
 
     Parameters
     ----------
@@ -268,14 +268,44 @@ def run_std_simulate(ns_run, estimator_list, n_simulate=None):
 
 
 def implementation_std(vals_std, vals_std_u, bs_std, bs_std_u, **kwargs):
-    """
-    Estimates implementation errors from the standard deviations of results
-    and of bootstrap values. See "Diagnostic tests for nested sampling
-    calculations" (Higson et al. 2018) for more details.
+    """Estimates varaition of results due to implementation-specific effects.
+    See "Diagnostic tests for nested sampling calculations" (Higson et al. 2018)
+    for more details.
 
-    Simulate errors dirstributions using the fact that (from central limit
-    theorem) our uncertainties on vals_std and bs_std are (approximately)
-    normally distributed.
+    Uncertainties on the output are calculated numerically using the fact that
+    (from central limit theorem) our uncertainties on vals_std and bs_std are
+    (approximately) normally distributed. This is needed as results from
+    standard error propagation techniques are not valid when the uncertainties
+    are not small compared to the result.
+
+    Parameters
+    ----------
+    vals_std: numpy array
+        Standard deviations of results from repeated calculations.
+    vals_std_u: numpy array
+        math:`1\sigma` uncertainties on vals_std_u.
+    bs_std: numpy array
+        Bootstrap error estimates. Each element should correspond to the same
+        element in vals_std.
+    bs_std_u: numpy array
+        math:`1\sigma` uncertainties on vals_std_u.
+    nsim: int, optional
+        Number of simulations to use to numerically calculate the uncertainties
+        on the estimated implementation-specific effects.
+    random_seed: int or None, optional
+        Numpy random seed. Use to get reproducible uncertainties on the output.
+
+    Returns
+    -------
+    imp_std: numpy array
+        Estimated standard deviation of results due to implementation-specific
+        effects.
+    imp_std_u: numpy array
+        math:`1\sigma` uncertainties on imp_std.
+    imp_frac: numpy array
+        imp_std as a fraction of vals_std.
+    imp_frac_u:
+        math:`1\sigma` uncertainties on imp_frac.
     """
     nsim = kwargs.pop('nsim', 1000000)
     random_seed = kwargs.pop('random_seed', 0)
@@ -306,18 +336,48 @@ def implementation_std(vals_std, vals_std_u, bs_std, bs_std_u, **kwargs):
 
 
 def run_thread_values(run, estimator_list):
[…]
