### 2018-09-11 ae87abf
improving docstrings
