### 2018-09-11 95e637a
minor tidying of docstrings
