+++ nestcheck/data_processing.py
@@ -1,13 +1,13 @@
 #!/usr/bin/env python
 """
 Functions for processing output files produced by nested sampling software.
-Currently compatable with MultiNest and PolyChord output.
+Currently compatable with ``MultiNest`` and ``PolyChord`` output.
 
 Nestcheck's diagnostics require infomation about the steps at which points were
 sampled in order to split nested sampling runs into their constituent threads
 (single live point runs). See "Sampling Errors In Nested Sampling Parameter
-Estimation" (Higson et al. 2017) for more details. *Producing these requires
-MultiNest >= v3.11 and PolyChord >= v1.13.*
+Estimation" (Higson et al. 2018) for more details. *Producing these requires
+``MultiNest`` >= v3.11 and ``PolyChord`` >= v1.13.*
 
 
 **nestested sampling run format**
@@ -50,8 +50,7 @@ import nestcheck.parallel_utils
 
 @nestcheck.io_utils.save_load_result
 def batch_process_data(file_roots, **kwargs):
-    """
-    Process output from many nested sampling runs in parallel with optional
+    """Process output from many nested sampling runs in parallel with optional
     error handling and caching.
 
     The result can be cached usin the 'save_name', 'save' and 'load' kwargs (by
@@ -120,8 +119,7 @@ def batch_process_data(file_roots, **kwargs):
 
 def process_error_helper(root, base_dir, process_func, errors_to_handle=(),
                          **func_kwargs):
-    """
-    Wrapper which applies process_func and handles some common errors so one
+    """Wrapper which applies process_func and handles some common errors so one
     bad run does not spoil the whole batch.
 
     Useful errors to handle include:
@@ -161,8 +159,7 @@ def process_error_helper(root, base_dir, process_func, errors_to_handle=(),
 
 def process_polychord_run(file_root, base_dir, process_stats_file=True,
                           **kwargs):
-    """
-    Loads data from a PolyChord run into the nestcheck dictionary format for
+    """Loads data from a PolyChord run into the nestcheck dictionary format for
     analysis.
 
     N.B. producing required output file containing information about the
@@ -206,8 +203,7 @@ def process_polychord_run(file_root, base_dir, process_stats_file=True,
 
 
 def process_multinest_run(file_root, base_dir, **kwargs):
-    """
-    Loads data from a MultiNest run into the nestcheck dictionary format for
+    """Loads data from a MultiNest run into the nestcheck dictionary format for
     analysis.
 
     N.B. producing required output file containing information about the
@@ -252,9 +248,8 @@ def process_multinest_run(file_root, base_dir, **kwargs):
 
 
 def process_dynesty_run(results):
-    """
-    Transforms results from a dynesty run into the nestcheck dictionary format
-    for analysis. This function has been tested with dynesty v9.2.0.
+    """Transforms results from a dynesty run into the nestcheck dictionary
+    format for analysis. This function has been tested with dynesty v9.2.0.
 
     Note that the nestcheck point weights and evidence will not be exactly
     the same as the dynesty ones as nestcheck calculates logX volumes more
@@ -305,8 +300,7 @@ def process_dynesty_run(results):
 
 
 def process_polychord_stats(file_root, base_dir):
-    """
-    Reads a PolyChord <root>.stats output file and returns the information
+    """Reads a PolyChord <root>.stats output file and returns the information
     contained in a dictionary.
 
     Parameters
@@ -364,8 +358,7 @@ def process_polychord_stats(file_root, base_dir):
 
 
 def process_samples_array(samples, **kwargs):
-    """
-    Convert an array of nested sampling dead and live points of the type
+    """Convert an array of nested sampling dead and live points of the type
[…]

[…]
