### 2018-09-11 6d271ba
further tidying of docstings
