### 2018-09-11 4ef3079
adding some more fancy sphynx formatting to the docsrings and updating paper references therein
