### 2018-09-11 5ade33f
adding more maths formating to docstrings
