### 2018-09-11 7015bd2
final update to docstring math rendering
