+++ docs/paper/paper.bib
@@ -130,13 +130,13 @@ year = {2018}
 @article{Higson2018perfectns,
 author = {Higson, Edward},
 doi = {10.5281/zenodo.1327591},
-title = {perfectns: perfect dynamic and standard nested sampling for spherically symmetric likelihoods and priors},
+title = {{perfectns: perfect dynamic and standard nested sampling for spherically symmetric likelihoods and priors}},
 year = {2018}
 }
 @misc{zenodonestcheck,
 author = {Higson, Edward},
 doi = {10.5281/zenodo.1329513},
-title = {nestcheck: error analysis, diagnostic tests and plots for nested sampling calculations},
+title = {{nestcheck: error analysis, diagnostic tests and plots for nested sampling calculations}},
 year = {2018}
 }
 @article{Higson2017a,
@@ -179,6 +179,6 @@ year = {2018}
 @article{Higson2018b,
 author = {Higson, Edward and Handley, William and Lasenby, Anthony and Hobson, Mike},
 journal = {in preparation},
-title = {{Bayesian Sparse Reconstruction: a brute-force approach to astronomical imaging and machine learning}},
+title = {{Bayesian sparse reconstruction: a brute-force approach to astronomical imaging and machine learning}},
 year = {2018}
 }
