### 2018-09-11 51ee47a
rewrite of data_processing module docstring to give more information about the information required in nested sampling output files and how new functions for processing nested sampling runs can be written
