+++ nestcheck/data_processing.py
@@ -1,35 +1,43 @@
 #!/usr/bin/env python
-"""
-Functions for processing output files produced by nested sampling software.
-Currently compatable with ``MultiNest`` and ``PolyChord`` output.
+r"""Module containing functions for loading and processing output files
+produced by nested sampling software.
 
-Nestcheck's diagnostics require infomation about the steps at which points were
-sampled in order to split nested sampling runs into their constituent threads
-(single live point runs). See "Sampling Errors In Nested Sampling Parameter
-Estimation" (Higson et al. 2018) for more details. *Producing these requires*
-``MultiNest`` *>= v3.11 and* ``PolyChord`` *>= v1.13.*
+Background: threads
+-------------------
 
+``nestcheck``'s error estimates and diagnostics rely on the decomposition
+of a nested sampling run into multiple runs, each with a single live point.
+We refer to these constituent single live point runs as *threads*.
+See "Sampling Errors In Nested Sampling Parameter Estimation" (Higson et
+al. 2018) for a detailed discussion, including an algorithm for dividing
+nested sampling runs into their constituent threads.
 
-**nestested sampling run format**
+Nested sampling run format
+--------------------------
 
-nestcheck stored nested sampling runs in a standard format as python
-dictionaries. For a run with nsamp samples, the keys are:
+``nestcheck`` stores nested sampling runs in a standard format as python
+dictionaries. For a run with :math:`n_\mathrm{samp}` samples, the keys are:
 
     logl: 1d numpy array
-        Log-likelihood values (floats) for each sample.
-        Shape is (nsamp,).
+        Loglikelihood values (floats) for each sample.
+        Shape is (:math:`n_\mathrm{samp}`,).
     thread_labels: 1d numpy array
-        Int representing which thread each point belongs to.
-        For some thread label k, the thread's start (birth) log-likelihood
-        and end log-likelihood are given by thread_min_max[k, :].
-        Shape is (nsamp,).
+        Integer label for each point representing which thread each point
+        belongs to.
+        Shape is (:math:`n_\mathrm{samp}`,).
+        For some thread label k, the thread's start (birth)
+        log-likelihood and end log-likelihood are given by
+        thread_min_max[k, :].
     thread_min_max: 2d numpy array
-        Shape is (# threads, 2).
-        Each row k contains min logl (birth contour) and max logl for
-        thread with thread label i.
+        Shape is (:math:`n_\mathrm{threads}`, 2).
+        Each row with index k contains the logl from within which the first
+        point in the thread with label k was sampled (the "birth contour") and
+        the logl of the final point in the thread.
+        The birth contour is -inf if the thread began by sampling from the
+        whole prior.
     theta: 2d numpy array
         Parameter values for samples - each row represents a sample.
-        Shape is (nsamp, d) where d is number of dimensions.
+        Shape is (:math:`n_\mathrm{samp}`, d) where d is number of dimensions.
     nlive_array: 1d numpy array
         Number of live points present between the previous point and
         this point.
@@ -37,6 +45,48 @@ dictionaries. For a run with nsamp samples, the keys are:
         Dict containing extra information about the run.
 
 Samples are arranged in ascending order of logl.
+
+Processing nested sampling software output
+------------------------------------------
+
+To process output files for a nested sampling run into the format described
+above, the following information is required:
+
+* Samples' loglikelihood values;
+* Samples' parameter values;
+* Information allowing decomposition into threads and identifying each thread's
+  birth contour (starting logl).
+
+The first two items are self-explanatory, but the latter is more challenging
+as it can take different formats and may not be provided by all nested sampling
+software packages.
+
+Sufficient information for thread decomposition and calculating the number of
[…]
