### 2018-09-11 eea2c12
renaming birth_inds_given_contours and threads_given_birth_inds to make their purposes clearer
