### 2018-09-11 748b223
minor affilation update
