### 2018-09-11 #916 @mattpitkin
@ejhigson would you be able to add a `__version__` parameter to the package, so that:

```python
print(nestcheck.__version__)
```

gives the current version information.
