### 2018-09-11 3b8f49d
added single-source __version__ property
