+++ nestcheck/__init__.py
@@ -1,1 +1,2 @@
 """initialise package directory."""
+from nestcheck._version import __version__
