### 2018-09-11 #916 @ejhigson
@mattpitkin thank you very much for your comments! I think I have addressed them - please let me know if there is anything else I should add or change.

> * add the DOI for the Matplotlib paper https://doi.org/10.1109/MCSE.2007.55
> * add the arXiv URL for the "Importance Nested Sampling..." paper https://arxiv.org/abs/1306.2144

Added.

> In the "**Calculating errors due to implementation-specific effects**" [section](https://nestcheck.readthedocs.io/en/latest/demos/quickstart_demo.html#Calculating-errors-due-to-implementation-specific-effects) could you state explicitly that this requires you to have multiple run (I know this is kind of covered in the next section, but it would be useful to have it).

Added.

> In the "**Tests for implementation specific effects using only 2 nested sampling runs**" [section](https://nestcheck.readthedocs.io/en/latest/demos/quickstart_demo.html#Tests-for-implementation-specific-effects-using-only-2-nested-sampling-runs) should the line:
>
> > As expected, the p-values are not particularly low nor are the KS distances particularly high - indicating there are significant implementation-specific effects for the simple 2-dimensional Gaussian
>
> actually say
>
> > As expected, the p-values are not particularly low nor are the KS distances particularly high - indicating there are **no** significant implementation-specific effects for the simple 2-dimensional Gaussian

You are correct - this is a (not insignificant) typo. Thank you for pointing it out! I have fix it.

> Would you be able to add some more information to the documentation about how to go about allowing samples from different nested sampling codes to be read in to `nestcheck`? I.e., what should a code output to allow you to use it with `nestcheck` through a custom `process_XXXX_runs` function? I see from the [`process_dynesty_runs`](https://github.com/ejhigson/nestcheck/blob/master/nestcheck/data_processing.py#L254) source that it seems to need values of the log likelihood, the actual sample values, the number of live points, and the sample ID - is the sample ID the order with which that sample was used for the nested sample integral (i.e., when that point "died")? What is a "thread" in this context?

I have added a guide to this in the module docstring of data_processing - let me know if this is ok/if it can be improved.

> @ejhigson would you be able to add a `__version__` parameter to the package, so that:
>
> ```python
> print(nestcheck.__version__)
> ```
>
> gives the current version information.

Added
