### 2018-09-12 ejhigson/nestcheck#1 @ziotom78
Unable to use results from `dynesty`
The `nestcheck` package provides support to load `dynesty` samples, but the following script produces an error using Python 3.6.6:
```python
# coding: utf-8
import dynesty
import numpy as np
from numpy import linalg
import math

ndim = 3  # number of dimensions
C = np.identity(ndim)  # set covariance to identity matrix
C[C==0] = 0.95 # set off-diagonal terms (strongly [Ccorrelated)
Cinv = linalg.inv(C)  # precision matrix
lnorm = -0.5 * (np.log(2 * np.pi) * ndim + np.log(linalg.det(C)))  # ln(normalization)

# 3-D correlated multivariate normal log-likelihood
def loglikelihood(x):
    """Multivariate normal log-likelihood."""
    return -0.5 * np.dot(x, np.dot(Cinv, x)) + lnorm

# prior transform
def prior_transform(u):
    """Transforms our unit cube samples `u` to a flat prior between -10. and 10. in each variable."""
    return 10. * (2. * u - 1.)

# initialize our nested sampler
sampler = dynesty.NestedSampler(loglikelihood, prior_transform, ndim, nlive=1000)
sampler.run_nested()
res = sampler.results

import nestcheck.data_processing
print('Dynesty version ', dynesty.__version__)
run = nestcheck.data_processing.process_dynesty_run(res)
```

The error is the following:
```
Traceback (most recent call last):
  File "mytest.py", line 30, in <module>
    run = nestcheck.data_processing.process_dynesty_run(res)
  File "/home/tomasi/nestcheck/lib/python3.6/site-packages/nestcheck/data_processing.py", line 301, in process_dynesty_run
    run = nestcheck.ns_run_utils.dict_given_run_array(samples, thread_min_max)
AttributeError: module 'nestcheck' has no attribute 'ns_run_utils'
```

Here are the versions of the package I'm using:
- dynesty 0.9.2
- NumPy 1.15.1
