### 2018-09-12 #916 @ziotom78
Hi @ejhigson , I have started testing your package, I found an issue and reported it here: [#1](https://github.com/ejhigson/nestcheck/issues/1).
