### 2018-09-12 ejhigson/nestcheck#1 @ejhigson
Whoops! @ziotom78 thanks for pointing this out - I had previously not noticed it due to importing ns_run_utils elsewhere. This should now be fixed in d2879b9, and your script works for me with this change. I will incorporate this fix into the next release along with any other changes from the JOSS review.
