### 2018-09-12 #916 @ziotom78
@ejhigson thanks for having fixed that problem so quickly! I have used the program and read the paper, for me everything looks ok.
