### 2018-09-14 6394766
updating to point to new version of diagnostic tests code for more examples
