### 2018-09-14 02d0ff7
reference updates for Bayesian spare reconstruction paper's arXiv identifier
