+++ docs/paper/paper.bib
@@ -172,13 +172,14 @@ arxivId = {1804.06406},
 author = {Higson, Edward and Handley, Will and Hobson, Mike and Lasenby, Anthony},
 eprint = {1804.06406},
 journal = {arXiv preprint arXiv:1804.06406},
-title = {{Diagnostic Tests for Nested Sampling Calculations}},
+title = {{nestcheck: diagnostic tests for nested sampling calculations}},
 url = {https://arxiv.org/abs/1804.06406},
 year = {2018}
 }
 @article{Higson2018b,
 author = {Higson, Edward and Handley, William and Lasenby, Anthony and Hobson, Mike},
-journal = {in preparation},
+journal = {arXiv preprint arXiv:1809.04598},
 title = {{Bayesian sparse reconstruction: a brute-force approach to astronomical imaging and machine learning}},
+url = {https://arxiv.org/abs/1809.04598},
 year = {2018}
 }
